+++ client/src/components/ShareModal/URLShareModal/URLShareModal.tsx
@@ -1,4 +1,4 @@
-import { UserInfo } from '../../types';
+import { UserInfo } from '../../../types';
 import { useState } from 'react';
 import URLShareSelectModal from '../URLShareSelectModal/URLShareSelectModal';
 import URLBannerShareModal from '../URLBannerShareModal/URLBannerShareModal';
@@ -13,9 +13,6 @@ enum ContentType {
   'BANNER',
 }
 
-// TODO: 색상버튼 값 바꿔야 됨
-// TODO: 색상 변경 버튼 클릭 시, 소스코드 및 이미지도 변경되어야함
-
 const URLShareModal = ({ userInfo, onClose }: URLShareModalProps) => {
   const [content, setContent] = useState<keyof typeof ContentType>('CHOICE');
 
@@ -23,19 +20,18 @@ const URLShareModal = ({ userInfo, onClose }: URLShareModalProps) => {
     setContent('BANNER');
   };
 
-  return (
-    <>
-      {content === 'CHOICE' ? (
-        <URLShareSelectModal
-          userInfo={userInfo}
-          onClose={onClose}
-          openBannerShareModal={openBannerShareModal}
-        />
-      ) : (
-        <URLBannerShareModal userInfo={userInfo} onClose={onClose} />
-      )}
-    </>
-  );
+  const urlShareModalObject = {
+    CHOICE: (
+      <URLShareSelectModal
+        userInfo={userInfo}
+        onClose={onClose}
+        openBannerShareModal={openBannerShareModal}
+      />
+    ),
+    BANNER: <URLBannerShareModal userInfo={userInfo} onClose={onClose} />,
+  };
+
+  return urlShareModalObject[content];
 };
 
 export default URLShareModal;
